+++ libs/dot-components/src/lib/components/list/List.styles.ts
@@ -1,0 +1,15 @@
+import styled, { css } from 'styled-components';
+import { List } from '@material-ui/core';
+
+export const rootClassName = 'dot-list';
+
+export const StyledList = styled(List)`
+  ${({ theme }) =>
+    css`
+      &.${rootClassName} {
+        &.dot-nested-list .dot-list-item {
+          padding-left: ${theme.spacing(4)}px;
+        }
+      }
+    `}
+` as typeof List;