+++ CHANGE_LOG.md
@@ -2,7 +2,9 @@
 
 ## Current
 
-- **ThemeProvider**: updated `breakpoints` in `theme` object to `xs: 0`, `sm: 720`, `md: 1024`, `lg: 1280`, `xl: 1920`,
+- **List** - added new component
+- **ListItem** - added new component
+- **ThemeProvider** updated `breakpoints` in `theme` object to `xs: 0`, `sm: 720`, `md: 1024`, `lg: 1280`, `xl: 1920`,
 
 ### Components